Functions

Functions

Normalised Hankel Function

InverseHankelFunction.hankelh1n — Function.

hankelh1n(ν, z₀ [,z])

Hankel function normalised at some z₀, H^{(1)}\nu(z)/H^{(1)}\nu(z₀), returns a function if z is not provided

source

InverseHankelFunction.HankelH1N — Type.

Struct to cache the value of the Hankel function of the first kind at z₀, can be called like a function after construction

source

Derivatives of Hankel function

InverseHankelFunction.diffhankelh1 — Function.

diffhankelh1(ν, z [, h[, hm1]])

Returns the first derivative of hankelh1 with respect to z. Can optionally pass in a precomputed H_ν(z) and H_{ν-1}(z).

source

InverseHankelFunction.diff2hankelh1 — Function.

diff2hankelh1(ν, z [, h[, hm1]])

Returns the first and second derivatives of hankelh1 with respect to z as tuple Can optionally pass in a precomputed H_ν(z) and H_{ν-1}(z).

source

InverseHankelFunction.diff3hankelh1 — Function.

diff3hankelh1(ν, z [, h[, hm1]])

Returns the first, second and third derivatives of hankelh1 with respect to z as tuple Can optionally pass in a precomputed H_ν(z) and H_{ν-1}(z).

source

Inverse of Normalised Hankel Function

InverseHankelFunction.invhankelh1n — Function.

invhankelh1n(ν::Integer, z₀::Number, h̄::Number)

Find z such that $H^{(1)}_\\nu(z)/H^{(1)}_\\nu(z₀) = \bar{h}$ for the branch continued from z₀ See also: hankelh1n

source

Missing docstring.

Missing docstring for invhankelh1n_sortedvec. Check Documenter's build log for details.

Missing docstring.

Missing docstring for diffinvhankelh1n_sortedvec. Check Documenter's build log for details.

Missing docstring.

Missing docstring for invhankelh1n_adaptive_solve. Check Documenter's build log for details.

Asymptotics of Hankel function

InverseHankelFunction.hankel_arg_asymptotic_scale — Function.

hankel_arg_asymptotic_scale(ν, z)

Determines largeness or smallness of argument to determine the validity of the asymptotic forms of the hankel functions.

source

InverseHankelFunction.small_arg_hankelh1 — Function.

small_arg_hankelh1(ν::Integer, z::Number) -> Complex

The small argument (z) asymptotic of hankelh1

source

InverseHankelFunction.inv_small_arg_hankelh1 — Function.

inv_small_arg_hankelh1(ν::Integer, h::Complex, b::Integer) -> Complex

The bth branch of the inverse of the small argument (z) asymptotic of hankelh1 Not valid for b > ν

source

InverseHankelFunction.large_arg_hankelh1 — Function.

large_arg_hankelh1(ν::Integer, z::Number) -> Complex

The large argument (z) asymptotic of hankelh1

source

InverseHankelFunction.inv_large_arg_hankelh1 — Function.

inv_large_arg_hankelh1(ν::Integer, h::Complex, b::Integer) -> Complex

The bth branch of the inverse of the large argument (z) asymptotic of hankelh1

source