Functions

Functions

Asymptotics

InverseHankelFunction.hankel_arg_asymptotic_scale — Function.

hankel_arg_asymptotic_scale(ν, z)

Determines largeness or smallness of argument to determine the validity of the asymptotic forms of the hankel functions.

source

InverseHankelFunction.small_arg_hankelh1 — Function.

small_arg_hankelh1(ν::Integer, z::Number) -> Complex

The small argument (z) asymptotic of hankelh1

source

InverseHankelFunction.inv_small_arg_hankelh1 — Function.

inv_small_arg_hankelh1(ν::Integer, h::Complex, b::Integer) -> Complex

The bth branch of the inverse of the small argument (z) asymptotic of hankelh1 Not valid for b > ν

source

InverseHankelFunction.large_arg_hankelh1 — Function.

large_arg_hankelh1(ν::Integer, z::Number) -> Complex

The large argument (z) asymptotic of hankelh1

source

InverseHankelFunction.inv_large_arg_hankelh1 — Function.

inv_large_arg_hankelh1(ν::Integer, h::Complex, b::Integer) -> Complex

The bth branch of the inverse of the large argument (z) asymptotic of hankelh1

source